Added network capacity
