+++ src/shared/constants.js
@@ -2,6 +2,8 @@ export const mainnetNodes = "https://raw.githubusercontent.com/Akashlytics/akash
 export const testnetNodes = "https://raw.githubusercontent.com/Akashlytics/akashlytics-deploy/master/testnet-nodes.json";
 export const edgenetNodes = "https://raw.githubusercontent.com/Akashlytics/akashlytics-deploy/master/edgenet-nodes.json";
 
+export const akashlyticsApi = "https://api.akashlytics.com/api"
+
 export const mainnetId = "mainnet";
 export const testnetId = "testnet";
 export const edgenetId = "edgenet";
